OneTensor, and changed template int dim to size_t dim

+++ 3bem/output.h
@@ -36,14 +36,14 @@ public:
     const int file_id;
 };
 
-template <int dim, typename T>
+template <size_t dim, typename T>
 void out_surface(Outputter& o, const Mesh<dim>& mesh,
                  const std::vector<T> data, int n_vars) {
     o.write_locations(mesh.facets.size(), dim * dim, mesh.facets.data());
     o.write_values(n_vars, reinterpret_vector<double>(data));
 }
 
-template <int dim, typename T> 
+template <size_t dim, typename T> 
 void out_volume(Outputter& o, const std::vector<Vec<double,dim>>& points,
                 const std::vector<T>& data, int n_vars) {
     o.write_locations(points.size(), dim, points.data());
